Rename + move directories
The initialisation file is now in an init folder; the experiment runner files are in a separate runner folder. A user who only uses the gui should only need to use files from the init folder, a user who runs the programs with the runners will use the runner folder as well. Also, fixed a spelling error in a comment.