Add reversed pulse to mFFTransSlice
The idea is that if you just pulse fast flux, you end up having offsets, due to reactive components in the signal chain. By performing the reverse of the pulse, you cancel out some of these (probably exact reverse is not optimal, but close enough). The pulse is not negative in this case, since the 'base' state is at some non-zero DAC value.